requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
DrissionPage==4.0.1
pandas==2.2.1
pyautogui==0.9.54
keyboard==0.13.5
pywinauto==0.6.9
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
requests==2.31.0
Pillow==11.1.0
pywin32==308
selenium==4.12.0
pytest==7.4.2
asgiref==3.7.2
asyncio==3.4.3 
scrapy==2.8.0
log==0.4.1
pytest-asyncio==0.21.0
mysql-connector-python==8.0.33
pyyaml==6.0
jieba==0.42.1
beautifulsoup4==4.12.2